add model

+++ src/main/java/com/karaoke/dao/UserDao.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.karaoke.model.User;
 
 @Repository
-public interface UserDao extends CrudRepository<User, Integer> {
+public interface UserDao extends CrudRepository<User, Long> {
 	
 	User findByUsername(String username);
